% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_freq_class.R
\name{add_freq_class}
\alias{add_freq_class}
\title{Convert percentages to frequency classes}
\usage{
add_freq_class(per_pres)
}
\arguments{
\item{per_pres}{Numeric. Percentage (i.e. 0-100)}
}
\value{
Factor with levels always, often, frequent, occasional, and infrequent
}
\description{
Convert percentages to frequency classes
}
\examples{
add_freq_class(c(2,20,40,60,100))
}
