% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_likelihood.R
\name{add_likelihood}
\alias{add_likelihood}
\title{Add likelihood based on IPCC thresholds}
\usage{
add_likelihood(df, col)
}
\arguments{
\item{df}{Dataframe with \code{col} raw values to assign to a likelihood class}

\item{col}{Column in \code{df} with raw values.}
}
\value{
\code{df} with extra columns from \code{envFunc::lulikelihood}.
}
\description{
Thresholds and defition based on the International Panel on Climate Change
\href{https://www.ipcc.ch/site/assets/uploads/2017/08/AR5_Uncertainty_Guidance_Note.pdf}{guidance note})
}
