% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{add_raster_cell}
\alias{add_raster_cell}
\title{Add cell to df with lat/long}
\usage{
add_raster_cell(df, ras = aoiRaster, x = "long", y = "lat", crsDf = 4326)
}
\arguments{
\item{df}{Dataframe with lat/long columns}

\item{ras}{Raster from which cell IDs will be extracted}

\item{x}{Character. Name of latitude column}

\item{y}{Character. Name of longitude column}
}
\value{

}
\description{
Add cell to df with lat/long
}
