% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_stamp.R
\name{add_time_stamp}
\alias{add_time_stamp}
\title{Add time stamp}
\usage{
add_time_stamp(obj)
}
\arguments{
\item{obj}{Any object to add a \code{ctime} attribute.}
}
\value{

}
\description{
Add a time stamp attribute to an object. From Stack Exchange Network
\href{https://stackoverflow.com/questions/5327555/object-creation-timestamp}{posts}
by
\href{https://stackoverflow.com/users/143305/dirk-eddelbuettel}{Dirk Eddelbuettel}.
}
