% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check packages}
\usage{
check_packages(packages, update_env = FALSE)
}
\arguments{
\item{packages}{Character. Packages to check}

\item{update_env}{Logical. Check for any updates to 'env' packages?}
}
\value{
Unique, sorted vector of package names
}
\description{
Check that packages are installed and install them if not. Optionally, make
sure that the most recent version of 'env' packages is installed.
}
\examples{
setup_packages(c("dplyr", "envFunc"))
}
