% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{clustering_summarise}
\alias{clustering_summarise}
\title{Basic summary of clusters}
\usage{
clustering_summarise(clustDf, groupName = "clust")
}
\arguments{
\item{clustDf}{dataframe. Contains a list column with output from cuttree in column groupName}

\item{groupName}{character. Name of column containing output from cuttree}
}
\value{
tibble
}
\description{
Basic summary of clusters
}
