% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{create_env}
\alias{create_env}
\title{Create dataframe of 'cells' and their associated env data}
\usage{
create_env(x, df, cores = 1, out_file = tempfile())
}
\arguments{
\item{x}{Raster* object.}

\item{df}{Dataframe. Must have a column called 'cell' that corresponds to
the cell numbers in x.}

\item{cores}{Numeric. Number of cores to use in snowfall::sfInit.}

\item{out_file}{Character. Path to save outputs.}
}
\value{
Dataframe with 'sites' column, plus columns equal to the length of x.
}
\description{
Create dataframe of 'cells' and their associated env data
}
