% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_test_func.R
\name{filter_test_func}
\alias{filter_test_func}
\title{Test rows against a function}
\usage{
filter_test_func(df, test_col = "path", test_func = terra::rast, ...)
}
\arguments{
\item{df}{Dataframe}

\item{test_col}{Character name of column in \code{df} to test against \code{test_func}.}

\item{test_func}{Name of function to test \code{test_col} with. \code{test_col} must be
the first argument to \code{test_func}.}

\item{...}{Other arguments passed to \code{test_func}.}
}
\value{
Filtered \code{df}.
}
\description{
Filter the rows of a dataframe that fail a function applied to a column.
}
