% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_name.R
\name{find_name}
\alias{find_name}
\title{Return a named object from within another object}
\usage{
find_name(haystack, needle)
}
\arguments{
\item{haystack}{Any object}

\item{needle}{A named object within \code{haystack}}
}
\value{
The object named \code{needle} or \code{NULL} if \code{needle} was not found within
\code{haystack}
}
\description{
from https://stackoverflow.com/questions/58400176/r-find-object-by-name-in-deeply-nested-list
}
\examples{
l <- list("a" = sample(letters, 10), "b" = list("c" = sample(letters, 2), "d" = sample(letters, 4)))
find_name(l, "d")
find_name(l, "b")
}
