% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bdbsa_taxa.R
\name{get_bdbsa_taxa}
\alias{get_bdbsa_taxa}
\title{Get taxa list from the Biological Database of South Australia (BDBSA)}
\usage{
get_bdbsa_taxa(
  source = "web",
  url =
    c("https://data.environment.sa.gov.au/Content/Publications/fauna-bdbsa-taxonomy.xlsx",
    "https://data.environment.sa.gov.au/Content/Publications/vascular-plants-bdbsa-taxonomy.xlsx",
    "https://data.environment.sa.gov.au/Content/Publications/other-taxonomic-groups-bdbsa-taxonomy.xlsx"),
  out_dir = tempdir(),
  standardise_taxonomy = FALSE,
  taxonomy = NULL,
  taxonomy_file = tempfile(),
  needed_ranks = c("species", "subspecies"),
  overrides = NULL,
  taxonomic_groups = c("fauna", "vascular-plants", "other"),
  filt_ranks = NULL,
  return_scientific = FALSE,
  current = TRUE
)
}
\arguments{
\item{source}{Character. Where to source the BDBSA taxonomy. Either 'web' (default) or 'oracle' (for internal DEW use only).}

\item{url}{If \code{source} = "web", a character vector of url paths to the relevant taxonomy xlsx links on https://data.environment.sa.gov.au.}

\item{out_dir}{Output directory for downloaded taxonomy xlsx files if \code{source} == "web".}

\item{standardise_taxonomy}{Logical. Standardise the taxa to the ALA taxonomic backbone via \code{envClean::make_taxonomy}?}

\item{taxonomy}{If !\code{standardise_taxonomy}, an optional alternative existing taxonomy list containing
\code{lutaxa} and \code{taxonomy} elements, as produced by \code{envClean::make_taxonomy}.}

\item{taxonomy_file}{If \code{standardise_taxonomy}, file path to save taxonomy results from \code{envClean::make_taxonomy}.}

\item{needed_ranks}{If \code{standardise_taxonomy}, character vector of taxonomic ranks required in the
returned list, as per \code{envClean::make_taxonomy}. Can be "all" or any combination of ranks from
envClean::lurank greater than or equal to subspecies. Each rank receives its own taxonomic list with
taxa binned up to that level, e.g. if "genus" then all species and subspecies are binned to genus,
but also returns all other taxa at levels higher than genus, i.e. family through to kingdom.}

\item{overrides}{Dataframe of taxa to override, as per \code{envClean::make_taxonomy}.}

\item{taxonomic_groups}{Optional character vector of broad taxonomic groups to filter the results, e.g. if only want fauna,
use taxonomic_groups = "fauna". Options are 'fauna', 'vascular-plants' & 'other' (for non-fauna and non-vascular plants).}

\item{filt_ranks}{If \code{standardise_taxonomy}, a character vector of taxonomic ranks to keep.
Note, these may be different to \code{needed_ranks}, which set the lowest taxonomic level to bin to, e.g.
with \code{needed_ranks} = "species" but without any \code{filt_ranks} this would return all taxa >= "species", but
with \code{filt_ranks} = "species" would only return species level taxa.}

\item{return_scientific}{Logcial. Return the scientific_name as per the ALA taxonomic backbone
(e.g. as opposed to species name in ALA which can be different)?}

\item{current}{Logical. Return only current BDBSA taxonomy? Applicable to \code{source} = "oracle" only.}
}
\value{
Dataframe with BDBSA original names, common names, NSX codes, indigenous status ('ind' field), and
standardised taxa names and returned ranks from ALA (if \code{standardise_taxonomy} or \code{taxonomy} is used).
}
\description{
Get taxa list from the Biological Database of South Australia (BDBSA)
}
