% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_dir.R
\name{get_env_dir}
\alias{get_env_dir}
\title{Return operating system independent 'env root' directory}
\usage{
get_env_dir(
  which_dir = NULL,
  linux_default = "/projects",
  windows_default = "H:"
)
}
\arguments{
\item{which_dir}{Character. Optional folder path below 'env root' directory}

\item{linux_default}{Character. If linux, where is 'env root'}

\item{windows_default}{Character. If windows, where is 'env root'}
}
\value{
Directory path ("fs_path", "character")
}
\description{
Return operating system independent 'env root' directory
}
