% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Return the most frequently occurring element of a vector}
\usage{
get_mode(x, ties = "sample", ...)
}
\arguments{
\item{x}{Vector}

\item{ties}{Character. If \code{ties = "sample"}, one value from the available
values will be randomly selected. Otherwise all ties are returned. Use
anything other than "sample" to return all ties.}

\item{...}{Passed to \code{base::table()}. Use for \code{NA} handling. Note default
handling is \code{useNA = "no"}, thus \code{NA} values will be ignored.}
}
\value{
Vector of the most frequent element(s) of \code{x}, with the same
class as \code{x}. If \code{ties = "sample"}, the length of the returned vector will be
1.
}
\description{
Ties are dealt with by either: randomly selecting from the available values,
thus always returning a vector of length 1; or returning all tied values,
thus potentially returning a vector of variable length.
}
\examples{
x <- sample(c(letters, rep(NA, 10)), 100, replace = TRUE)
get_mode(x)
x <- x |> factor(levels = letters)
get_mode(x)
get_mode(x, useNA = "always")

x <- c(10, 10, 20, 20)
get_mode(x)
get_mode(x, ties = "blah")
}
