% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_or_make}
\alias{get_or_make}
\title{Make an object if it is not available from provided file}
\usage{
get_or_make(saveFile, forceNew = FALSE, fn, ...)
}
\arguments{
\item{saveFile}{Character. Path to file. This will be loaded to object if it
exists}

\item{forceNew}{Logical. If TRUE, run function irrespective of existence of
saveFile. Otherwise, only run fn if saveFile does not exist.}

\item{fn}{Function. Function to make the object if it saveFile does not
exist}

\item{...}{Arguments to function.}
}
\value{
Dataframe. Also, file if it did not previously exist.
}
\description{
Make an object if it is not available from provided file
}
