% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scale.R
\name{get_scale}
\alias{get_scale}
\title{Get env scales from a yaml}
\usage{
get_scale(element = "default", scales = "settings/scales.yaml")
}
\arguments{
\item{element}{Name of element in \code{scales} for which to extract the
scale.}

\item{scales}{Path to yaml containing the scales relevant to a project.}
}
\value{
Named, nested list describing \code{element}
}
\description{
Requires a scale named 'default' in \code{scales}. All other elements describe
change from 'default'. \code{NULL} values in default are maintained. Set a
non-default list element to \code{NULL} to remove it from default.
}
\examples{
extent <- list(extent = list(geo = "sa_br_dissolve", temp = "P10Y", blah = NULL))
grain <- list(grain = list(geo = "90", temp = "P1Y"))
settings <- list(default = c(extent, grain))
change <- list(change = list(extent = list(blah = NULL), grain = list(geo = "30")))
settings <- c(settings, change)
get_scale(scales = settings)
get_scale(element = "change", scales = settings)

}
