% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit_env.R
\name{git_commit_env}
\alias{git_commit_env}
\title{Add, commit and push all current changes to github}
\usage{
git_commit_env(comment)
}
\arguments{
\item{comment}{Character commit comment.}
}
\value{
Side effect of adding, committing and pushing any changes in project.
}
\description{
Add, commit and push all current changes to github
}
