% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain_to_bin.R
\name{grain_to_bin}
\alias{grain_to_bin}
\title{Extract bin information from the grain element of a scales yaml}
\usage{
grain_to_bin(
  scales_file = "settings/scales.yaml",
  scales_element = basename(here::here()),
  add = c("cell_lat", "cell_long"),
  remove = c("res_", "grain_time", "taxonomic")
)
}
\arguments{
\item{scales_file}{Character. Path to the \code{scales.yaml}}

\item{scales_element}{Character or numeric. Element within \code{scales_file} with
the grain from which to extract bin information}

\item{add}{Character. Any extra bin columns that cannot be extracted from the
grain}

\item{remove}{Character. Any regular expressions to exclude from the bin
columns}
}
\description{
A \code{bin} defines the columns in an \code{env} workflow that collectively define
the grain around which cleaning processes are aimed. This function is
intended to explicitly link the grain in a \code{scales.yaml} with the bins used
in the workflow.
}
