% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_ctime.R
\name{last_ctime}
\alias{last_ctime}
\title{Object with the latest \code{ctime} attribute}
\usage{
last_ctime(prefix = "bio_")
}
\arguments{
\item{prefix}{}
}
\value{
Character. Object name with latest \code{ctime} attribute.
}
\description{
Object with the latest \code{ctime} attribute
}
