% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulikelihood}
\alias{lulikelihood}
\title{Dataframe of likelihood thresholds and definitions}
\format{
A data frame with 9 rows and
7 variables:
\describe{
\item{likelihood}{Factor. Definitions.}

\item{maxVal}{Double. Maximum proportion to fall in likelihood category}

\item{range}{Factor. Result of
\code{cut(lulikelihood$maxVal, breaks = c(0,lulikelihood$maxVal))}}

\item{loose}{Factor with levels:
+, 0 and -}

\item{very}{Factor with levels:
++, +, 0, - and --}

\item{extreme}{Factor with levels:
+++, ++, +, 0, -, -- and ---}

\item{exceptional}{Factor with levels:
++++, +++, ++, +, 0, -, --, --- and ----}

...
}
}
\source{
\url{https://www.ipcc.ch/site/assets/uploads/2017/08/AR5_Uncertainty_Guidance_Note.pdf}
}
\usage{
lulikelihood
}
\description{
Following the intergovernmental panel on climate change \href{https://www.ipcc.ch/site/assets/uploads/2017/08/AR5_Uncertainty_Guidance_Note.pdf}{IPCC}.
}
\keyword{datasets}
