% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulsa}
\alias{lulsa}
\title{Lookup for Landscapes South Australia regions}
\format{
A data frame with 9 rows and 9 variables:
\describe{
\item{LSA}{Character. LSA code}
\item{REGION}{Character. LSA Region name}
\item{LSA_region}{Character. Same as REGION}
\item{Zone}{Character. Agricultural or Arid zone}
\item{System}{Character. High- or Low-rainfaill or Arid}
\item{R,G,B,A}{Numeric. Red, Green, Blue and Alpha values to create colour
values for \code{LSA}.}
...
}
}
\usage{
lulsa
}
\description{
Lookup for Landscapes South Australia regions
}
\keyword{datasets}
