% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aoi.R
\name{make_aoi}
\alias{make_aoi}
\title{Make a region (area) of interest}
\usage{
make_aoi(
  polygons,
  filt_col = NULL,
  filt_level = NULL,
  simplify = TRUE,
  buffer = FALSE,
  bbox = FALSE,
  clip = NULL,
  clip_buf = 100,
  buf_crs = 7845,
  out_crs = 4326
)
}
\arguments{
\item{polygons}{sf}

\item{filt_col}{Character name of column in \code{polygons} to filter on (or \code{NULL}
for no filter)}

\item{filt_level}{Level(s) of \code{filt_col} to filter}

\item{buffer}{Numeric (or \code{FALSE}). Distance to buffer in units of \code{buf_crs}}

\item{bbox}{Logical. Return a bounding box around the result?}

\item{clip}{sf. If not null, this will be used to clip back the original
\code{polygons} +/- filtered +/- buffered +/- bbox. CHECK both CRS and \code{clip_buf}}

\item{clip_buf}{Numeric (or \code{FALSE}). Distance to buffer \code{clip} in units of \code{clip}}

\item{buf_crs}{\href{https://epsg.io/}{EPSG} code to use when buffering}

\item{out_crs}{\href{https://epsg.io/}{EPSG} code for coordinate system of
result}
}
\description{
Make a region (area) of interest
}
