% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_clusters}
\alias{make_clusters}
\title{Apply clustering algorithms}
\usage{
make_clusters(
  df,
  methodsDf = clustMethod,
  sppCol = "Taxa",
  siteCol = "list",
  numCol = "p",
  groups = 2:100,
  minTaxaCount = 1
)
}
\arguments{
\item{df}{Dataframe. Needs the 'Cols' outlined below.}

\item{methodsDf}{Dataframe.}

\item{minTaxaCount}{}
}
\value{

}
\description{
Apply clustering algorithms
}
