% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_epochs.R
\name{make_epochs}
\alias{make_epochs}
\title{Generate a tibble of epochs.}
\usage{
make_epochs(
  min_year = 1987,
  max_year = 2020,
  epoch_step = 5,
  epoch_overlap = TRUE
)
}
\arguments{
\item{min_year}{Minimum year in epochs}

\item{max_year}{Maximum year in epochs}

\item{epoch_step}{Numeric. How many years in an epoch?}

\item{epoch_overlap}{Logical. Should epochs overlap by one year? i.e.
\code{epoch_step = 10} with \code{epoch_overlap = TRUE} gives, say, 2000-2010 and
2010-2020 whereas \code{epoch_overlap = FALSE} gives 2000-2009 and
2010-2019.}
}
\value{
Tibble with columns
\describe{
\item{start}{First year of epoch}
\item{end}{Last year of epoch}
\item{years}{List column containing the years for the epoch}
\item{epoch}{Last two digits of first year and last year in epoch.
e.g. 10-19.}
\item{data}{List column of full path of raster files to summarise.}
}
}
\description{
Generate a tibble of epochs.
}
\examples{
now <- as.numeric(format(Sys.Date(), "\%Y"))
make_epochs(now - 30, now, 10)
make_epochs(now - 30, now, 10, F)

}
