% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metric_df.R
\name{make_metric_df}
\alias{make_metric_df}
\title{Use a set of (continuous) columns to choose a good set of rows}
\usage{
make_metric_df(
  df,
  mets_df = tibble::tibble(metric = "av_clust_size", high_good = TRUE, summary_mets =
    TRUE, weight = 1),
  context = c("method", "groups"),
  mets_col = "summary_mets",
  summarise_method = median,
  top_thresh = 0.25,
  best_thresh = 5,
  scale = lifecycle::deprecated(),
  level = lifecycle::deprecated()
)
}
\arguments{
\item{df}{Dataframe with columns over which to find good rows}

\item{mets_df}{Dataframe mapping the name of possible metrics to cases
(columns) in which to use that metric.}

\item{context}{Character. Name of columns in df that define context.}

\item{mets_col}{Character. Name of \code{mets_df} column to use in this instance.}

\item{summarise_method}{Character. Name of method to use in summarising if
there is more than one row per context.}

\item{top_thresh}{Numeric specifying the proportion of rows considered 'top'.}

\item{best_thresh}{Numeric specifying the absolute number of rows considered
'best'.}
}
\description{
Use a set of (continuous) columns to choose a good set of rows
}
