% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metric_plot.R
\name{make_metric_plot}
\alias{make_metric_plot}
\title{Plot the results from diagnostic_df}
\usage{
make_metric_plot(
  metric_df,
  group_col = "groups",
  facet_col = "method",
  label = "metric",
  value = "scale",
  display_all = FALSE
)
}
\arguments{
\item{metric_df}{Dataframe with results from call to \code{metrics_df}.}

\item{group_col}{Character. Name of column with numeric data for x-axis.}

\item{facet_col}{Character. Optional name of column to facet on (along with label).}

\item{label}{Character label for the diagnostics (choose another column from diagnosticDF).}

\item{value}{Character. Name of column with values to plot.}

\item{display_all}{Display all diagnostics or only those used to select best.}
}
\value{

}
\description{
Plot the results from diagnostic_df
}
