% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make package workflow}
\usage{
make_package(do_commit = FALSE, commit_message)
}
\arguments{
\item{commit_message}{}
}
\value{

}
\description{
document, install, knit README.Rmd, build_site, commit
}
