% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seasons.R
\name{make_seasons}
\alias{make_seasons}
\title{Make a list of data frames for months and seasons}
\usage{
make_seasons(
  start_year,
  end_year,
  seasons = NULL,
  dec_adjust = 1,
  include_all = TRUE
)
}
\arguments{
\item{start_year, end_year}{Numeric. Year of first and last season}

\item{seasons}{Dataframe mapping each month to a season. Default (\code{NULL})
will use south hemisphere seasons. Needs columns \code{mon} (numeric 1:12), \code{month}
(character name to use for months) and \code{season} (character name of season to
use for that month)`}

\item{dec_adjust}{Logical. Adjust the year in which December is used? If
\code{TRUE} (default) December will be included in the following year's first
season (i.e. December 1999 will be included in, say, summer 2000)}

\item{include_all}{Logical. Include a 'season' that is 'all' (and starts at
calendar year).}
}
\value{
List with elements
item{luseasons}{Dataframe. Lookup used to translate months to seasons. Equal
to \code{seasons} if that was provided by \code{seasons} argument}
item{months}{Dataframe for each month from \code{start_year} to \code{end_year},
including start and end dates}
item{seasons}{Dataframe for each season from \code{start_year} to \code{end_year},
including start and end dates}
}
\description{
Make a list of data frames for months and seasons
}
