% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{make_targets}
\alias{make_targets}
\title{Make targets}
\usage{
make_targets(settings_context_file = "settings/setup.yaml")
}
\arguments{
\item{settings_context_file}{}
}
\value{
list of 'projects' each with elements 'script' and 'store'. Saves
_targets.yaml
}
\description{
Runs \code{targets::tar_source()}, reads the context from \code{settings_context_file},
and creates _targets.yaml from any files matching the pattern
\code{"^\\\\d{3}_.*\\\\.R$"}.
}
