% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_env_out.R
\name{name_env_out}
\alias{name_env_out}
\title{Use meta data to name or parse output paths}
\usage{
name_env_out(
  x,
  context_defn = c("vector", "filt_col", "level", "buffer", "res"),
  parse = FALSE,
  fill_null = FALSE,
  x_null = 2,
  ...
)
}
\arguments{
\item{x}{Either dataframe with path(s) to parse (in column \code{path}), character
vector of path(s) to search, or named list object.}

\item{context_defn}{Character vector of meta data names (in order)}

\item{parse}{Logical. If \code{FALSE} (default) an \code{out_dir} path will be returned rather
than parsed. Assumes the appropriate names can be found in \code{x}.}

\item{fill_null}{Logical. If \code{TRUE}, will fill up to \code{x_null} definitions
with NULL (and issue a warning).}

\item{x_null}{Numeric. Even if \code{fill_null} is \code{TRUE}, if there are more than
\code{x_null} missing definitions, an error will be thrown.}

\item{...}{Not used}
}
\value{
If \code{!parse}, tibble with columns:
\describe{
\item{vector}{Vector layer originally used to define area of interest}
\item{filt_col}{Column name from vector to filter to define area of interest}
\item{level}{Level(s) of filt_col originally filtered to define area of interest}
\item{buffer}{Any buffer around area of interest}
\item{res}{Resolution of output raster in units of rasters crs}
\item{path}{Full (relative) path including \code{file_type}.}
}

If \code{!parse}, \code{df} with additional column \code{out_dir}
}
\description{
Either create file path for saving outputs or parse the meta data from a path
}
