% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_df.R
\name{project_df}
\alias{project_df}
\title{Transform the coordinate columns of a dataframe}
\usage{
project_df(
  df,
  use_aoi,
  x = "long",
  y = "lat",
  new_x = "x",
  new_y = "y",
  crs_from = "epsg:4326",
  crs_to = "epsg:8059"
)
}
\arguments{
\item{df}{Dataframe. Needs coordinate columns}

\item{x, y}{Character. Name of column with x and coords}

\item{new_x, new_y}{Character. Name to give the transformed x and columns}

\item{crs_from, crs_to}{Passed to the \code{from} and \code{to} arguments of
\code{sf::sf_project()}}
}
\value{
Dataframe with additional, projected, x and y columns
}
\description{
Uses \code{sf::sf_project()} internally
}
