% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{prop_mem}
\alias{prop_mem}
\title{Proportion of current memory usage}
\usage{
prop_mem()
}
\value{
Numeric. Proportion
}
\description{
Edited from various posts on the \href{https://stackoverflow.com/questions/27788968/how-would-one-check-the-system-memory-available-using-r-on-a-windows-machine}{Stack Exchange Network}.
}
\examples{
if(.Platform$OS.type == "windows") prop_mem()
}
