% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quibble.R
\name{quibble}
\alias{quibble}
\title{Make a wide, single row, data frame of quantiles (percentiles)}
\usage{
quibble(x, q = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{x}{Numeric. Values passed to \code{stats::quantile} argument \code{x}}

\item{q}{Numeric. Proportions passed to \code{stats::quantile} argument \code{probs}}

\item{...}{Passed to \code{stats::quantile}}
}
\value{
Single row data frame. Names are percentiles equivalent to \code{q}
values.
}
\description{
Edited from the blog post https://www.tidyverse.org/blog/2020/03/dplyr-1-0-0-summarise/
}
\examples{
x <- rnorm(100)

quibble(x)
}
