% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the scrips in a project}
\usage{
run(from = 0, to, skips = NULL)
}
\arguments{
\item{from}{numeric. minimum script to run}

\item{to}{numeric. maximum script to run}

\item{skips}{character. anything to grep from the file names that you do not want to run}
}
\value{
side effect of sourcing the scripts
}
\description{
Assumes the file names of the scripts to run start with four-digits that order the files. i.e. 0010_setup.R will run before 0020_clean.R
}
