% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_targets.R
\name{run_targets}
\alias{run_targets}
\title{Run targets}
\usage{
run_targets(settings_context_file = "settings/setup.yaml")
}
\arguments{
\item{settings_context_file}{}
}
\value{
as per \code{targets::tar_make()}
}
\description{
Runs \code{targets::tar_source()}, reads the context from \code{settings_context_file},
creates _targets.yaml from any files matching the pattern
\code{"^\\\\d{3}_.*\\\\.R$"} and then runs all the targets.
}
