% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_transform_df.R
\name{sp_transform_df}
\alias{sp_transform_df}
\title{Transform the coordinates in a dataframe}
\usage{
sp_transform_df(
  df,
  to_crs,
  df_crs = 4283,
  x = "long",
  y = "lat",
  remove = TRUE
)
}
\arguments{
\item{df}{Dataframe with x and y coordinates to transform.}

\item{df_crs, to_crs}{Anything that will be recognised by the \code{crs} argument
of \code{sf::st_as_sf()} and \code{sf::st_transform()}.}

\item{x, y}{Character name of column in \code{df} with x and y coordinates.}

\item{remove}{Logical. If true, original coordinate columns (\code{x} and \code{y})
will be removed.}
}
\value{
Dataframe with added columns \code{X} and \code{Y} representing the transformed
coordinates.
}
\description{
Transform the coordinates in a dataframe
}
