% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_dir.R
\name{store_dir}
\alias{store_dir}
\title{Generate a path to a storage directory}
\usage{
store_dir(set_list, base_dir = fs::path("..", "..", "out"))
}
\arguments{
\item{set_list}{Named, nested list. Passed to \code{set_list} argument of
\code{envFunc::name_env_out()}}

\item{base_dir}{Character. Path to outputs store.}
}
\value{
Full path to a store directory for a project
}
\description{
Generate a path to a storage directory
}
\examples{
# a list of settings
settings <- list(extent = list(polygons = "sa_ibrasub_xn"
                               , filt_col = NULL
                               , filt_level = NULL
                               , buffer = 0
                               , temp_ext = "P50Y"
                               )
                 , grain = list(x = 90
                                , y = 90
                                , z = "P50Y"
                                , taxonomic = "species"
                                )
                 , aoi = list(polygons = "lsa"
                              , filt_col = "LSA"
                              , filt_level = "GA"
                              , buffer = 0
                              )
                 )

# generate the full path to a store
out_dir <- store_dir(settings)

out_dir
}
