% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{summarise_env}
\alias{summarise_env}
\title{Environmental summary statistics}
\usage{
summarise_env(env_df, context, luenv_df = NULL, trans = "transform")
}
\arguments{
\item{env_df}{Data frame of environmental values.}

\item{context}{Character name of column(s) in \code{env_df} that define the context.}

\item{luenv_df}{Data frame with information about each layer.}
}
\value{
Data frame of summary values for each layer, particularly \code{units} and
\code{transform} to convert raster layer values to \code{units}.
}
\description{
Given a data frame of environmental values, calculate summary statistics
}
