% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_long_df.R
\name{summarise_long_df}
\alias{summarise_long_df}
\title{Summarise a long data frame with consistent set of summary functions}
\usage{
summarise_long_df(
  long_df,
  name_col = "name",
  value_col = "value",
  group_cols = NULL,
  qs = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{long_df}{Dataframe in long format}

\item{name_col}{Character. Group values in \code{long_df} by this column name. Can
be \code{NULL} to summarise the entire \code{long_df}}

\item{value_col}{Character. Name of column in \code{long_df} with the values to
summarise}

\item{group_cols}{Character. Any extra columns in \code{long_df} to group by.}

\item{qs}{Numeric. Passed to the \code{q} argument of \code{stats::quantile()} (via
\code{envFunc::quibble()})}
}
\value{
tibble of summarised values with one row for: each unique value in
\code{name_col}; and, if supplied, the combination with \code{group_cols}.
}
\description{
Summary functions are: count; max; min; mean; quantiles at each value of
\code{qs}; count of \code{NA} values; and proportion of \code{NA} values.
}
