% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_label.R
\name{taxa_label}
\alias{taxa_label}
\title{Create a label for a species}
\usage{
taxa_label(common = NULL, taxa = NULL, records = NULL, what_for = "plot")
}
\arguments{
\item{common}{Character. Common name for species}

\item{taxa}{Character. Scientific name for species}

\item{records}{Numeric. Optional number of records.}

\item{what_for}{"plot" or something else. If plot will be formatted for
plotting otherwise formatted for markdown.}
}
\description{
Default is 'Common (\emph{Scientific}). Possibly including \verb{n = }
if \code{records} is supplied. Formatted for either ggplot (default
\code{what_for =plot}) or markdown.
}
\examples{
cat(taxa_label("Black-chinned Honeyeater", "Melithreptus gularis"))
cat(taxa_label(taxa = "Melithreptus gularis"))
cat(taxa_label("Black-chinned Honeyeater", records = 3))
cat(taxa_label("Black-chinned Honeyeater", "Melithreptus gularis", what_for = "md"))
cat(taxa_label("Black-chinned Honeyeater", "Melithreptus gularis", what_for = "md", records = 3))
}
