% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cores.R
\name{use_cores}
\alias{use_cores}
\title{Returns a suggested number of cores}
\usage{
use_cores(prop_max_detected = 0.5, absolute_max = 128)
}
\arguments{
\item{prop_max_detected}{What proportion of \code{parallel::detect_cores()} to
use?}

\item{absolute_max}{Integer. Absolute maximum number of cores to use.}
}
\value{
Integer
}
\description{
Provide the output of \code{use_cores()} to, say, \code{crew::crew_controller_local()}
\code{workers} argument.
}
\examples{
use_cores()
}
