% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{add_raster_cell}
\alias{add_raster_cell}
\title{Add cell to df with lat/long}
\usage{
add_raster_cell(ras, df, x = "long", y = "lat", crsdf = 4326)
}
\arguments{
\item{ras}{Raster object with cell numbers to extract}

\item{df}{Dataframe with lat/long columns}

\item{x}{Character. Name of latitude column}

\item{y}{Character. Name of longitude column}

\item{crsdf}{Single length vector. What crs are x and y?}
}
\value{
df with additional column 'cell' of cell numbers from ras.
}
\description{
Add cell to df with lat/long
}
