% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dir.R
\name{clean_dir}
\alias{clean_dir}
\title{Empty a directory}
\usage{
clean_dir(dir_to_clean, skip = "old")
}
\arguments{
\item{dir_to_clean}{Path to directory to 'clean'}

\item{skip}{Character. Any strings that identify files to skip (identified
via \code{base::grep()})}
}
\value{
Returns \code{NULL}. Any files in \code{dir_to_clean} are moved.
}
\description{
Any files in \code{dir_to_clean} are moved to the new directory
\code{dir_to_clean/old/moved_\%Y\%m\%d_\%H\%M\%S}
}
