% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_up.R
\name{first_up}
\alias{first_up}
\title{Change the first letter of a string to capital}
\usage{
first_up(string)
}
\arguments{
\item{string}{}
}
\value{

}
\description{
Doesn't alter the capitalisation of any other characters, just the first one.
}
\examples{
first_up("DEW")
first_up("water from space")
first_up("geoscience Australia")
}
