% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make package workflow}
\usage{
make_package(do_commit = FALSE, m)
}
\arguments{
\item{do_commit}{Logical. Commit to github?}

\item{m}{Character. Commit message to include.}
}
\value{
Called for side effect of running \code{\link[devtools:document]{devtools::document()}},
\code{\link[devtools:install]{devtools::install()}}, \link[knitr:knit("README.Rmd")]{knitr::knit("README.Rmd")}, [pkgdown::build_site()[
and, optionally, \code{\link[=git_commit_env]{git_commit_env()}}.
}
\description{
document, install, knit README.Rmd, build_site, commit
}
