% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make package workflow}
\usage{
make_package(m = NULL, ...)
}
\arguments{
\item{m}{Character. Commit message to include. If \code{NULL}, then no commit.}

\item{...}{Passed to \code{pkgdown::build_site()}}
}
\value{
Called for side effect of: making any data in data-raw; and then
running \code{devtools::document()}, \code{knitr::knit()},
"README.Rmd", \code{pkgdown::clean_site()}, \code{pkgdown::build_site()}, and,
optionally, \code{envFunc::git_commit_env()}.
}
\description{
make any data-raw; document; knit README.Rmd; clean_site; build_site; commit
}
