% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make package workflow}
\usage{
make_package(do_commit = FALSE, m, ...)
}
\arguments{
\item{do_commit}{Logical. Commit to github?}

\item{m}{Character. Commit message to include}

\item{...}{Passed to \code{pkgdown::build_site()}}
}
\value{
Called for side effect of running \code{devtools::document()},
\code{devtools::install()}, \code{knitr::knit()} "README.Rmd", \code{pkgdown::build_site()}
and, optionally, \code{envFunc::git_commit_env()}.
}
\description{
All now wrapped in if(FALSE) so doesn't do anything. Best run as line by line
manually.
}
\details{
document, install, knit README.Rmd, build_site, commit

WARNING. Runs rm(list = ls()), leaving only arguments to this function.
}
