% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_system.R
\name{monitor_system}
\alias{monitor_system}
\title{Monitor system resources}
\usage{
monitor_system(monitor_df = NULL, plot = TRUE, plot_time = 1/24)
}
\arguments{
\item{monitor_df}{NULL or results from previous call to envFunc::monitor_system()}
}
\value{
Dataframe
}
\description{
A simple version of windows task manager
}
\examples{
  df <- monitor_system()

  counter <- 1

  while(counter < 3) {

    counter <- counter + 1

    Sys.sleep(1)

    df <- monitor_system(df)

  }
}
