% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert a numeric to its corresponding english character.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
'spelled out' (in english) x.
}
\description{
From https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r
}
