% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert a numeric to its corresponding english character.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
Character. 'spelled out' (in english) x.
}
\description{
Edited from the \href{https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r}{Github}
file by \href{https://github.com/ateucher}{ateucher}.
}
\examples{
numbers2words(10)
numbers2words(floor(sample(1:1000,1)))
}
