% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vec_to_sentence}
\alias{vec_to_sentence}
\title{Vector to sentence}
\usage{
vec_to_sentence(x, sep = ",", end = "and")
}
\arguments{
\item{x}{Character. Vector to collapse to a sentence.}

\item{sep}{Character. Separator between all except last.}

\item{end}{Character. Last separator.}
}
\value{

}
\description{
Turn a vector into a comma separated list of values with a penultimate 'and'
or other separator.
}
\examples{
x <- c("apples", "bannanas", "pears", "grapes")
vec_to_sentence(x)
vec_to_sentence(x,end = "&")
}
